+++ EnchantmentsPlusMinus/src/org/vivi/eps/util/economy/Economy.java
@@ -10,7 +10,7 @@ public interface Economy {
 	 * @param amount The amount changed. Can be negative.
 	 * @return Returns the amount the player has after.
 	 */
-	public Integer changeBalance(String playername, Integer amount);
+	public double changeBalance(String playername, double amount);
 	
 	/**Changes the specified player's balance.
 	 * 
@@ -18,7 +18,7 @@ public interface Economy {
 	 * @param amount The amount changed. Can be negative.
 	 * @return Returns the amount the player has after.
 	 */
-	public Integer changeBalance(Player player, Integer amount);
+	public double changeBalance(Player player, double amount);
 	
 	/**Sets the specified player's balance.
 	 * 
@@ -26,7 +26,7 @@ public interface Economy {
 	 * @param value The amount to be set to.
 	 * @return Returns the balance the player has after.
 	 */
-	public Integer setBalance(String playername, Integer value);
+	public double setBalance(String playername, double value);
 	
 	/**Sets the specified player's balance.
 	 * 
@@ -34,19 +34,19 @@ public interface Economy {
 	 * @param value The amount to be set to.
 	 * @return Returns the balance the player has after.
 	 */
-	public Integer setBalance(Player player, Integer value);
+	public double setBalance(Player player, double value);
 	
 	/**Returns the balance of the specified player.
 	 * 
 	 * @param player The player to get from
 	 * @return Returns the balance of the specified player.
 	 */
-	public Integer getBalance(Player player);
+	public double getBalance(Player player);
 	
 	/**Returns the balance of the specified player.
 	 * 
 	 * @param playername The player to get from
 	 * @return Returns the balance of the specified player.
 	 */
-	public Integer getBalance(String playername);
+	public double getBalance(String playername);
 }
